Push metadata helpers from coreapi down to coreunix so that tests there can pick up the functions also.
