+++ core/coreapi/object.go
@@ -11,6 +11,7 @@ import (
 	"io"
 	"io/ioutil"
 
+	"github.com/TRON-US/go-btfs/core/coreunix"
 	"github.com/TRON-US/go-btfs/dagutils"
 	"github.com/TRON-US/go-btfs/pin"
 
@@ -146,7 +147,7 @@ func (api *ObjectAPI) Data(ctx context.Context, path ipath.Path, unixfs bool, me
 	if unixfs {
 		ds := api.core().Dag()
 
-		pData, metaData, err := getDataForUserAndMeta(ctx, pbnd, ds)
+		pData, metaData, err := coreunix.GetDataForUserAndMeta(ctx, pbnd, ds)
 		if err != nil {
 			return nil, nil, err
 		}
@@ -159,81 +160,6 @@ func (api *ObjectAPI) Data(ctx context.Context, path ipath.Path, unixfs bool, me
 	} else {
 		return bytes.NewReader(pbnd.Data()), nil, nil
 	}
-}
-
-func (api *ObjectAPI) MetaDataMap(ctx context.Context, path ipath.Path) (map[string]interface{}, error) {
-	metaData, err := api.MetaData(ctx, path)
-	if err != nil {
-		return nil, err
-	}
-
-	jmap := make(map[string]interface{})
-	err = json.Unmarshal(metaData, &jmap)
-	if err != nil {
-		return nil, err
-	}
-
-	return jmap, nil
-}
-
-func (api *ObjectAPI) MetaData(ctx context.Context, path ipath.Path) ([]byte, error) {
-	nd, err := api.core().ResolveNode(ctx, path)
-	if err != nil {
-		return nil, err
-	}
-
-	_, ok := nd.(*dag.ProtoNode)
-	if !ok {
-		return nil, dag.ErrNotProtobuf
-	}
-
-	ds := api.core().Dag()
-
-	metaData, err := getMetaData(ctx, nd, ds)
-	if err != nil {
-		return nil, err
-	}
-
-	return metaData, nil
-}
-
-// Return user data, metadata in byte array, and error.
-// Note that this function assumes, if token metadata exists inside
-// the DAG topped by the given 'node', the 'node' has metadata root
-// as first child, user data root as second child.
-func getDataForUserAndMeta(ctx context.Context, nd ipld.Node, ds ipld.DAGService) ([]byte, []byte, error) {
-	n := nd.(*dag.ProtoNode)
-
-	fsType, err := ft.GetFSType(n)
-	if err != nil {
-		return nil, nil, err
-	}
-
-	if ft.TTokenMeta == fsType {
-		return nil, n.Data(), nil
-	}
-
-	// Return user data and metadata if first child is of type TTokenMeta.
-	if nd.Links() != nil && len(nd.Links()) >= 2 {
-		childen, err := ft.GetChildrenForDagWithMeta(ctx, nd, ds)
-		if err != nil {
-			return nil, nil, err
-		}
-		if childen == nil {
-			return nil, n.Data(), nil
-		}
-		return childen[1].(*dag.ProtoNode).Data(), childen[0].(*dag.ProtoNode).Data(), nil
-	}
-
-	return n.Data(), nil, nil
-}
-
-// Return metadata under the given 'nd' node
-// if 'nd' has metadata root node aa its first child.
-func getMetaData(ctx context.Context, nd ipld.Node, ds ipld.DAGService) ([]byte, error) {
-
-	_, metadata, err := getDataForUserAndMeta(ctx, nd, ds)
-	return metadata, err
 }
 
 func (api *ObjectAPI) Links(ctx context.Context, path ipath.Path) ([]*ipld.Link, error) {
